refactor(alt-text): dedupe endpoint test fixtures into shared harness
Extract the duplicated buildPluginConfig and fake-request builders from
the four endpoint test files into test/support/endpointHarness.ts. The
shared buildPluginConfig takes overrides so each test varies only the one
facet it exercises (locales, maxBulkGenerateIds, a recording resolver),
and buildEndpointRequest records the findByID/update Local API calls.
