+++ alt-text/test/endpointAccessControl.test.ts
@@ -3,10 +3,11 @@ import assert from 'node:assert/strict'
 import { Forbidden } from 'payload'
 import { describe, test } from 'vitest'
 
-import type { PayloadRequest } from 'payload'
-
-import type { AltTextPluginConfig } from '../src/types/AltTextPluginConfig.ts'
-
+import {
+  buildEndpointRequest as buildRequest,
+  buildPluginConfig,
+  testUser as user,
+} from './support/endpointHarness.ts'
 import { bulkGenerateAltTextsEndpoint } from '../src/endpoints/bulkGenerateAltTexts.ts'
 import { generateAltTextEndpoint } from '../src/endpoints/generateAltText.ts'
 
@@ -17,62 +18,6 @@ import { generateAltTextEndpoint } from '../src/endpoints/generateAltText.ts'
  * able to touch. Every Local API call the endpoints make must run under the
  * requesting user's access (`overrideAccess: false` + `user: req.user`).
  */
-
-const user = { id: 'low-priv-user', email: 'user@example.com', role: 'user' }
-
-type LocalApiCall = Record<string, unknown>
-
-function buildPluginConfig(): AltTextPluginConfig {
-  return {
-    access: ({ req }) => !!req.user,
-    collections: [{ slug: 'media', mimeTypes: ['image/*'] }],
-    enabled: true,
-    getImageThumbnail: () => 'https://example.com/thumb.png',
-    healthCheck: true,
-    healthCheckAccess: ({ req }) => !!req.user,
-    locale: 'en',
-    locales: [],
-    maxBulkGenerateConcurrency: 1,
-    resolver: {
-      key: 'mock',
-      resolve: async () => ({
-        success: true,
-        result: { altText: 'generated alt', keywords: ['a', 'b'] },
-      }),
-      resolveBulk: async () => ({
-        success: true,
-        results: { en: { altText: 'generated alt', keywords: ['a', 'b'] } },
-      }),
-    },
-  }
-}
-
-function buildRequest(body: unknown): {
-  findByIDCalls: LocalApiCall[]
-  req: PayloadRequest
-  updateCalls: LocalApiCall[]
-} {
-  const findByIDCalls: LocalApiCall[] = []
-  const updateCalls: LocalApiCall[] = []
-
-  const req = {
-    json: async () => body,
-    payload: {
-      config: { custom: { altTextPluginConfig: buildPluginConfig() } },
-      findByID: async (args: LocalApiCall) => {
-        findByIDCalls.push(args)
-        return { id: args.id, filename: 'photo.png', mimeType: 'image/png' }
-      },
-      update: async (args: LocalApiCall) => {
-        updateCalls.push(args)
-        return { id: args.id }
-      },
-    },
-    user,
-  } as unknown as PayloadRequest
-
-  return { findByIDCalls, req, updateCalls }
-}
 
 describe('generate endpoint access control', () => {
   test('reads the document under the requesting user, not with overridden access', async () => {
